requirements.txt
streamlit
mistralai
openai 
python-dotenv 